+++ validat1.cpp
@@ -240,7 +240,7 @@ bool TestOS_RNG()
 		time_t t = time(NULL), t1 = 0;
 
 		// check that it doesn't take too long to generate a reasonable amount of randomness
-		while (total < 16 && (t1 < 10 || total*8 > t1))
+		while (total < 16 && (t1 < 10 || total*8 > (unsigned long)t1))
 		{
 			test.Pump(1);
 			total += 1;
